CVProject/pyproject.toml
[tool.poetry]
name = "cvproject"
version = "0.1.0"
description = "A simple CV listing and PDF/email service"
authors = ["axlrose023 <sloboda282@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
Django = "5.2.4"
reportlab = ">=4.4.3,<5.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
psycopg = {extras = ["binary"], version = ">=3.2.9,<4.0.0"}
celery = ">=5.5.3,<6.0.0"
redis = "^6.2.0"
openai = "^1.97.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"